requirements.txt
pyyaml
instamob
